fix clang compile error in method cChunkMap::DoExplosionAt

+++ src/ChunkMap.cpp
@@ -1820,7 +1820,7 @@ void cChunkMap::DoExplosionAt(double a_ExplosionSize, double a_BlockX, double a_
 				float EntityExposure = a_Entity.GetExplosionExposureRate(ExplosionPos, ExplosionSizeInt);
 
 				// Exposure reduced by armor
-				EntityExposure = EntityExposure * (1.0 - a_Entity.GetEnchantmentBlastKnockbackReduce());
+				EntityExposure = static_cast<float>(EntityExposure * (1.0 - a_Entity.GetEnchantmentBlastKnockbackReduce()));
 
 				double Impact = (1 - (Length / ExplosionSizeInt / 2)) * EntityExposure;
